Display weather  information

+++ WeatherDemo/Models/WeatherDataModel.swift
@@ -9,12 +9,12 @@
 import UIKit
 
 class WeatherDataModel: NSObject {
-    var minTemp: Double = 0.0
-    var maxTemp: Double = 0.0
-    var currentTemp: Double = 0.0
-    var humidity: Double = 0.0
+    var minTemp: Int = 0
+    var maxTemp: Int = 0
+    var currentTemp: Int = 0
+    var humidity: Int = 0
     var windSpeed: Double = 0.0
-    var windDeg: Double = 0.0
+    var windDeg: Int = 0
     
     // MARK: - Intializer
     override init() {
@@ -24,15 +24,15 @@ class WeatherDataModel: NSObject {
         self.init()
         let main: [String: Any] = (attributes["main"] as? [String: Any])!
         
-        minTemp                  = main["temp_min"] as? Double ?? 0.0
-        maxTemp                  = main["temp_max"] as? Double ?? 0.0
-        currentTemp              = main["temp"] as? Double ?? 0.0
-        humidity                 = main["humidity"] as? Double ?? 0.0
+        minTemp                  = main["temp_min"] as? Int ?? 0
+        maxTemp                  = main["temp_max"] as? Int ?? 0
+        currentTemp              = main["temp"] as? Int ?? 0
+        humidity                 = main["humidity"] as? Int ?? 0
         
         
         let wind: [String: Any] = (attributes["wind"] as? [String: Any])!
         windSpeed                 = wind["speed"] as? Double ?? 0.0
-        windDeg                  = wind["deg"] as? Double ?? 0.0
+        windDeg                  = wind["deg"] as? Int ?? 0
         
     }
